Rebase

+++ examples/delete/delete.ino
@@ -15,24 +15,26 @@
  ****************************************************/
 
 #include <Adafruit_Fingerprint.h>
-#include <SoftwareSerial.h>
-
-uint8_t getFingerprintEnroll(uint8_t id);
-
-// Software serial for when you dont have a hardware serial port
-// pin #2 is IN from sensor (GREEN wire)
-// pin #3 is OUT from arduino  (WHITE wire)
-// On Leonardo/Micro/Yun, use pins 8 & 9. On Mega, just grab a hardware serialport 
-SoftwareSerial mySerial(2, 3);
-Adafruit_Fingerprint finger = Adafruit_Fingerprint(&mySerial);
 
 // On Leonardo/Micro or others with hardware serial, use those! #0 is green wire, #1 is white
-//Adafruit_Fingerprint finger = Adafruit_Fingerprint(&Serial1);
+// uncomment this line:
+// #define mySerial Serial1
+
+// For UNO and others without hardware serial, we must use software serial...
+// pin #2 is IN from sensor (GREEN wire)
+// pin #3 is OUT from arduino  (WHITE wire)
+// comment these two lines if using hardware serial
+#include <SoftwareSerial.h>
+SoftwareSerial mySerial(2, 3);
+
+Adafruit_Fingerprint finger = Adafruit_Fingerprint(&mySerial);
 
 void setup()  
 {
   Serial.begin(9600);
-  Serial.println("Delete Finger");
+  while (!Serial);  // For Yun/Leo/Micro/Zero/...
+  delay(100);
+  Serial.println("\n\nDelete Finger");
 
   // set the data rate for the sensor serial port
   finger.begin(57600);
@@ -45,21 +47,26 @@ void setup()
   }
 }
 
+
+uint8_t readnumber(void) {
+  uint8_t num = 0;
+  
+  while (num == 0) {
+    while (! Serial.available());
+    num = Serial.parseInt();
+  }
+  return num;
+}
+
 void loop()                     // run over and over again
 {
-  while (!Serial);  // For Yun/Leo/Micro/Zero/...
-  delay(500);
-
-  Serial.println("Type in the ID # you want delete...");
-  uint8_t id = 0;
-  while (true) {
-    while (! Serial.available());
-    char c = Serial.read();
-    if (! isdigit(c)) break;
-    id *= 10;
-    id += c - '0';
+  Serial.println("Please type in the ID # (from 1 to 127) you want to delete...");
+  uint8_t id = readnumber();
+  if (id == 0) {// ID #0 not allowed, try again!
+     return;
   }
-  Serial.print("deleting ID #");
+
+  Serial.print("Deleting ID #");
   Serial.println(id);
   
   deleteFingerprint(id);
